Extract snapshot resolution into agent/document_view/snapshot.rs
